requirements.txt
langchain-openai>=0.3.0
langchain-tavily>=0.1.0
langchain-core>=0.3.0
openai>=1.0.0
pandas>=2.0.0
streamlit>=1.32.0
python-dotenv>=1.0.0
